+++ lib/services/payments/payments_provider.g.dart
@@ -53,4 +53,4 @@ final class PaymentsServiceProvider
   }
 }
 
-String _$paymentsServiceHash() => r'538e4d386eea1b852056a9a3cbc7fdcbebcf1166';
+String _$paymentsServiceHash() => r'abea181fa9b990e28717f47946a6cd337f884984';